pyproject.toml
[tool.poetry]
name = "openhands-dynamic-agents"
version = "0.1.0"
description = "Dynamic agent generation module for OpenHands"
authors = ["OpenHands Team <openhands@all-hands.dev>"]
readme = "README.md"
packages = [{include = "openhands_dynamic_agents", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
openhands = "^0.1.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
pyyaml = "^6.0.1"
click = "^8.1.7"
plotly = "^5.18.0"
pandas = "^2.1.3"
jinja2 = "^3.1.2"
aiofiles = "^23.2.1"
websockets = "^12.0"

[tool.poetry.scripts]
agents = "openhands_dynamic_agents.cli:main"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"